카메라 진동 구현

+++ Assets/Scripts/Object/InteractionObject.cs
@@ -4,13 +4,13 @@ public abstract class InteractionObject : MonoBehaviour
 {
     [Header("Interaction Object")]
     [SerializeField]
-    protected int maxHP = 100;
-    protected int currentHP;
+    protected float maxHP = 100;
+    protected float currentHP;
 
     private void Awake()
     {
         currentHP = maxHP;
     }
 
-    public abstract void TakeDamage(int damage);
+    public abstract void TakeDamage(float damage);
 }
